An Anarch Manifesto
", The" is really the only "prefix as a suffix" we should have. It's painful enough to maintain honestly. Postfixing the "An" here is a very hard to detect issue for any tool relying on the CSV.

eg. even the generated images have a mistake in them because of it. https://static.krcg.org/card/anarchmanifestoan.jpg